Handle metadata fields consistently.
The calibration fields used for equivalence were not set well.
Instead of passing text strings around, we should use a
camera/detector objects and grab the fields we care about in a single
way.  These are now parameters of the updateMetadata method, which
should now force all calibrations to have consistent field usage.

This also defines the CALIB_ID.